Fixed signed/unsigned errors in unit tests

+++ tests/q/src/execution_context.cpp
@@ -59,7 +59,7 @@ TEST( execution_context, thread_pool_dispatcher )
 	q::all( std::move( promises ), tpqu )
 	.then( [ num, &val, tasks, bd, tpctx ]( )
 	{
-		EXPECT_EQ( val.load( ), num + tasks );
+		EXPECT_EQ( static_cast< std::size_t >( val.load( ) ), num + tasks );
 		bd->terminate( q::termination::linger );
 		tpctx->dispatcher( )->terminate( q::termination::linger );
 	}, qu );
